+++ app/packages/core/src/components/Actions/types.ts
@@ -1,0 +1,5 @@
+import type { AdaptiveMenuItemComponentPropsType } from "@fiftyone/components";
+
+export type ActionProps = {
+  adaptiveMenuItemProps?: AdaptiveMenuItemComponentPropsType;
+};
