+++ app/packages/core/src/components/Grid/Actions/DynamicGroup/index.tsx
@@ -1,10 +1,8 @@
-import {
-  AdaptiveMenuItemComponentPropsType,
-  PillButton,
-} from "@fiftyone/components";
+import { PillButton } from "@fiftyone/components";
 import LoadingDots from "@fiftyone/components/src/components/Loading/LoadingDots";
 import * as fos from "@fiftyone/state";
 import MergeIcon from "@mui/icons-material/Merge";
+import type { MouseEvent } from "react";
 import React, {
   useCallback,
   useEffect,
@@ -14,22 +12,19 @@ import React, {
 } from "react";
 import useMeasure from "react-use-measure";
 import { useRecoilValue } from "recoil";
+import type { ActionProps } from "../../../Actions/types";
+import { ActionDiv, getStringAndNumberProps } from "../../../Actions/utils";
 import DynamicGroup from "./DynamicGroup";
-import { ActionDiv, getStringAndNumberProps } from "./utils";
 
 const DYNAMIC_GROUP_PILL_BUTTON_ID = "dynamic-group-pill-button";
 
-export const DynamicGroupAction = ({
-  adaptiveMenuItemProps,
-}: {
-  adaptiveMenuItemProps: AdaptiveMenuItemComponentPropsType;
-}) => {
+export default ({ adaptiveMenuItemProps }: ActionProps) => {
   const { refresh } = adaptiveMenuItemProps;
   const [open, setOpen] = useState(false);
   const [isProcessing, setIsProcessing] = useState(false);
 
   const [measureRef] = useMeasure();
-  const ref = useRef();
+  const ref = useRef<HTMLDivElement>(null);
 
   const pillComponent = useMemo(() => {
     if (isProcessing) {
@@ -41,12 +36,16 @@ export const DynamicGroupAction = ({
   }, [isProcessing]);
 
   useEffect(() => {
+    /** refresh */
+    isProcessing;
+    /** refresh */
+
     refresh?.();
   }, [isProcessing, refresh]);
 
   const isDynamicGroupViewStageActive = useRecoilValue(fos.isDynamicGroup);
 
-  const onCloseHandler = useCallback((e?: React.MouseEvent<Element>) => {
+  const onCloseHandler = useCallback((e?: MouseEvent<Element>) => {
     const pillButtonElement = document.getElementById(
       DYNAMIC_GROUP_PILL_BUTTON_ID
     );
