+++ app/packages/core/src/components/Grid/Spotlight.tsx
@@ -1,0 +1,123 @@
+import styles from "./Grid.module.css";
+
+import Spotlight from "@fiftyone/spotlight";
+import * as fos from "@fiftyone/state";
+import React, { useState } from "react";
+import { useRecoilValue } from "recoil";
+import { useMemoOne } from "use-memo-one";
+import { v4 as uuid } from "uuid";
+import { useSyncLabelsRenderingStatus } from "../../hooks";
+import {
+  gridAutosizing,
+  gridCrop,
+  gridSpacing,
+  maxGridItemsSizeBytes,
+  pageParameters,
+} from "./recoil";
+import useEscape from "./useEscape";
+import useEvents from "./useEvents";
+import useLabelVisibility from "./useLabelVisibility";
+import useLookerCache from "./useLookerCache";
+import useRecords from "./useRecords";
+import useRefreshers from "./useRefreshers";
+import useRenderer from "./useRenderer";
+import useResize from "./useResize";
+import useScrollLocation from "./useScrollLocation";
+import useSpotlightPager from "./useSpotlightPager";
+import useUpdates from "./useUpdates";
+import useZoomSetting from "./useZoomSetting";
+
+const MAX_INSTANCES = 5151;
+const MAX_ROWS = 5151;
+
+function Spotlight() {
+  const id = useMemoOne(() => uuid(), []);
+  const pixels = useMemoOne(() => uuid(), []);
+  const spacing = useRecoilValue(gridSpacing);
+  const { pageReset, reset } = useRefreshers();
+  const [resizing, setResizing] = useState(false);
+  const zoom = useZoomSetting();
+
+  useSyncLabelsRenderingStatus();
+
+  const records = useRecords(pageReset);
+
+  // divide by two, half for the hidden cache and half for max shown
+  const maxBytes = useRecoilValue(maxGridItemsSizeBytes) / 2;
+  const cache = useLookerCache({
+    maxHiddenItems: MAX_INSTANCES,
+    maxHiddenItemsSizeBytes: maxBytes,
+    reset,
+    ...useLabelVisibility(),
+  });
+
+  const { page, store } = useSpotlightPager({
+    clearRecords: reset,
+    pageSelector: pageParameters,
+    records,
+    zoomSelector: gridCrop,
+  });
+
+  const { getFontSize, lookerOptions, renderer } = useRenderer({
+    cache,
+    id,
+    records,
+    store,
+  });
+  const { get, set } = useScrollLocation(pageReset);
+
+  const setSample = fos.useExpandSample(store);
+  const autosizing = useRecoilValue(gridAutosizing);
+
+  const spotlight = useMemoOne(() => {
+    /** SPOTLIGHT REFRESHER */
+    reset;
+    /** SPOTLIGHT REFRESHER */
+
+    if (resizing) {
+      return undefined;
+    }
+
+    cache.freeze();
+
+    return new Spotlight<number, fos.Sample>({
+      ...get(),
+      ...renderer,
+
+      maxRows: MAX_ROWS,
+      maxItemsSizeBytes: autosizing ? maxBytes : undefined,
+      scrollbar: true,
+      spacing,
+
+      get: (next) => page(next),
+      onItemClick: setSample,
+      rowAspectRatioThreshold: zoom,
+    });
+  }, [
+    cache,
+    autosizing,
+    get,
+    maxBytes,
+    page,
+    renderer,
+    reset,
+    resizing,
+    setSample,
+    spacing,
+    zoom,
+  ]);
+
+  useEscape();
+  useEvents({ id, cache, pixels, resizing, set, spotlight });
+  useUpdates({ cache, getFontSize, options: lookerOptions, spotlight });
+  useResize(id, setResizing);
+
+  return (
+    <div className={styles.gridContainer}>
+      <div id={id} className={styles.spotlightGrid} data-cy="fo-grid" />
+      <div id={pixels} className={styles.fallingPixels} />
+    </div>
+  );
+}
+
+export default React.memo(Spotlight);
