+++ app/packages/state/src/recoil/atoms.ts
@@ -288,7 +288,7 @@ export const similarityParameters = (() => {
 
   return graphQLSyncFragmentAtom<
     datasetFragment$key,
-    State.SortBySimilarityParameters | null
+    State.SortBySimilarityParameters
   >(
     {
       fragments: [datasetFragment],
